fix: EACCESS error for package `devalue`
In the assumption this error is caused because the built files are not fing independent and
they require some external package that must have been embeded during the build step.
or I am stupid. that I went through the bundled files to see what was wrong after googling and searching for some time.